impove readability of code by classifying/merging to particular packages
