+++ app/src/androidTest/java/com/github/se/signify/ui/screens/profile/FriendsListScreenTest.kt
@@ -1,10 +1,9 @@
-package com.github.se.signify.ui.screens
+package com.github.se.signify.ui.screens.profile
 
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createComposeRule
 import com.github.se.signify.ui.navigation.NavigationActions
-import com.github.se.signify.ui.screens.profile.FriendsListScreen
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
